utils_metro without test
since metro is using access limit control from cloudfare. It seems that it will block the ip for certain time if it opens 6 pages in a certain timeframe.
For now it will not go through each item's page to get description and ingredient.
For future update, it can use proxy or web crawler ditributed at different locations to get item detail.